Started refactoring to increase code reuse
