+++ src/core/merkle/IMerkleTreeConfig.ts
@@ -1,0 +1,4 @@
+export interface IMerkleTreeConfig {
+    version: number;
+    layerCount: number;
+}
