+++ src/core/merkle/LamportGenerator.spec.ts
