+++ src/core/merkle/LamportGenerator.ts
