+++ src/core/merkle/MerkleTreeHelper.spec.ts
@@ -1,0 +1,24 @@
+import { getConfigStorageKey, getLayerStorageKeys } from "./MerkleTreeHelper";
+
+describe("MerkleTreeHelper", () => {
+    it("should return correct config storage keys", () => {
+        let dummyWallet = {
+            id: "WALLET_ID"
+        };
+
+        expect(getConfigStorageKey(<any>dummyWallet)).toBe("WALLET_ID-config");
+    });
+
+    it("should return correct layer storage keys", () => {
+        let dummyWallet = {
+            id: "WALLET_ID"
+        };
+
+        expect(getLayerStorageKeys(<any>dummyWallet, 4)).toEqual([
+            "WALLET_ID-layer-0",
+            "WALLET_ID-layer-1",
+            "WALLET_ID-layer-2",
+            "WALLET_ID-layer-3"
+        ]);
+    });
+})
