+++ src/core/merkle/MerkleTreeVersion.ts
@@ -1,0 +1,1 @@
+export const MERKLE_TREE_VERSION = 1;
