+++ src/models/ILocalWallet.ts
@@ -1,9 +1,8 @@
 import { IWallet } from "./IWallet";
 import { IKeyStore } from "./IKeyStore";
-import { MerkleTree } from "../merkle/MerkleTree";
+import { MerkleTree } from "../core/merkle/MerkleTree";
 
 export interface ILocalWallet extends IWallet {
     type: "local";
     keyStore: IKeyStore;
-    merkleTree?: MerkleTree;
 }
