+++ src/services/merkle-tree-service/merkle-tree-service.ts
@@ -1,10 +1,15 @@
 import { Injectable } from "@angular/core";
 import { IWallet } from "../../models/IWallet";
-import { MerkleTree, IMerkleTreeConfig } from "../../merkle/MerkleTree";
+import { MerkleTree } from "../../core/merkle/MerkleTree";
 import { KeyStoreService } from "../key-store-service/key-store-service";
 import { Storage } from "@ionic/storage";
 import { ILocalWallet } from "../../models/ILocalWallet";
 import { Platform } from "ionic-angular/platform/platform";
+import { MerkleTreeBuilder } from "../../core/merkle/MerkleTreeBuilder";
+import { MerkleTreeSerializer } from "../../core/merkle/MerkleTreeSerializer";
+import { MerkleTreeDeserializer } from "../../core/merkle/MerkleTreeDeserializer";
+import { getConfigStorageKey, getLayerStorageKeys } from "../../core/merkle/MerkleTreeHelper";
+import { IMerkleTreeConfig } from "../../core/merkle/IMerkleTreeConfig";
 
 export interface IMerkleTreeService {
     generate(wallet: IWallet, password: string, progressUpdate?: (progress: number) => void): Promise<void>;
@@ -17,6 +22,10 @@ export interface IMerkleTreeService {
 @Injectable()
 export class MerkleTreeService implements IMerkleTreeService {
     private cache: {[index: string]: MerkleTree} = {};
+
+    private merkleTreeBuilder = new MerkleTreeBuilder();
+    private merkleTreeSerializer = new MerkleTreeSerializer();
+    private merkleTreeDeserializer = new MerkleTreeDeserializer();
 
     constructor(private keyStoreService: KeyStoreService,
                 private storage: Storage,
@@ -31,13 +40,13 @@ export class MerkleTreeService implements IMerkleTreeService {
             return Promise.reject("Could not decrypt keystore");
 
         // Start generating the Merkle Tree
-        return MerkleTree.generate(privateKey, 14, this.platform, progressUpdate).then(
+        return this.merkleTreeBuilder.generate(privateKey, 14, this.platform.is("android"), progressUpdate).then(
             (merkleTree) => {
                 // Cache Merkle Tree
                 this.cache[wallet.id] = merkleTree;
                 
                 // Store Merkle Tree on disk
-                return merkleTree.serialize(wallet.id, this.storage, this.keyStoreService, password);
+                return this.merkleTreeSerializer.serialize(merkleTree, wallet, this.storage, this.keyStoreService, password);
             }
         );
     }
@@ -49,7 +58,7 @@ export class MerkleTreeService implements IMerkleTreeService {
         }
         else {
             // Read from disk and then cache
-            return MerkleTree.fromDisk(wallet, this.storage, this.keyStoreService, password).then(
+            return this.merkleTreeDeserializer.fromDisk(wallet, this.storage, this.keyStoreService, password).then(
                 (merkleTree) => {
                     this.cache[wallet.id] = merkleTree;
 
@@ -64,12 +73,12 @@ export class MerkleTreeService implements IMerkleTreeService {
         delete this.cache[wallet.id];
 
         // Remove from disk
-        return this.storage.get(MerkleTree.getConfigStorageKey(wallet)).then(
+        return this.storage.get(getConfigStorageKey(wallet)).then(
             (config: IMerkleTreeConfig) => {
                 if(!config)
                     return Promise.resolve();
 
-                let layerKeys = MerkleTree.getLayerStorageKeys(wallet, config.layerCount);
+                let layerKeys = getLayerStorageKeys(wallet, config.layerCount);
 
                 let promises: Promise<void>[] = [];
 
@@ -90,7 +99,7 @@ export class MerkleTreeService implements IMerkleTreeService {
 
                 // Finaly remove the config
                 promises.push(
-                    this.storage.remove(MerkleTree.getConfigStorageKey(wallet))
+                    this.storage.remove(getConfigStorageKey(wallet))
                 );
 
                 return Promise.all(promises).then<void>();
